sleftv-bridge: try to get the table conversion logic "right"
In some cases, we can predict the output type (e.g. when directly
encoded in the tables), but when conversions are involved, we never
know which method will be called: it migth depend on the state of
the system, for example: `1 - (1..1)` will return a different
type of object depending on whether a current ring is defined.
So when there is ambiguity (more than 1 possible result type),
we don't choose until runtime what type to output.